+++ nolds/__init__.py
@@ -1,3 +1,3 @@
 from .measures import lyap_r, lyap_e, sampen, hurst_rs, corr_dim, dfa, \
     fbm, binary_n, logarithmic_n, logarithmic_r
-from .datasets import brown72
+from .datasets import brown72, tent_map, logistic_map